introduction and fix of dumb modifying-a-copy bug. Also many variables are now signed to better detect  overflow (but also this shouldn't happen since many were expanded to 64 bits

+++ cpp-games/src/Attacker.h
@@ -5,14 +5,14 @@
 
 class Attacker : public Player {
     public:
-        uint id;
+        uint64_t id;
 
         static double inequality_ratio;
         
         static long long  a_init;
-        static long long  attacker_iter_sum; // how much the attackers have cumulatively gained or lost this round
+        static int64_t  attacker_iter_sum; // how much the attackers have cumulatively gained or lost this round
         static std::vector<unsigned long long> cumulative_assets; // running total of all attackers' assets
-        static long long  current_sum_assets; // sum total of all class instances
+        static int64_t current_sum_assets; // sum total of all class instances
 
         static double estimated_current_defender_posture_mean;
 
@@ -26,8 +26,8 @@ class Attacker : public Player {
         static void reset();
 
         Attacker(int id_in,  Params &p);
-        void gain(uint32_t gain) override;
-        void lose(uint32_t loss) override;
+        void gain(int64_t gain) override;
+        void lose(int64_t loss) override;
 
         // To re-estimate current defender params
         static void perform_market_analysis(std::vector<Defender> &defenders);
