+++ cpp-games/src/Defender.h
@@ -9,14 +9,14 @@ class Defender : public Player {
     public:
         static double estimated_probability_of_attack;
 
-        static unsigned long long d_init;
-        static unsigned long long defender_iter_sum;
+        static int64_t d_init;
+        static int64_t defender_iter_sum;
         static std::vector<unsigned long long> cumulative_assets; // running total of all defenders' assets
-        static unsigned long long current_sum_assets; // sum total of all class instances
-        static unsigned long long sum_recovery_costs;
-        static unsigned long long policiesPurchased;
-        static unsigned long long defensesPurchased;
-        static unsigned long long sum_security_investments;
+        static int64_t current_sum_assets; // sum total of all class instances
+        static int64_t sum_recovery_costs;
+        static int64_t policiesPurchased;
+        static int64_t defensesPurchased;
+        static int64_t sum_security_investments;
         // static unsigned int doNothing; // TODO implement? what is this for? maybe delete
 
         static double ransom_b0;
@@ -24,7 +24,7 @@ class Defender : public Player {
         static double recovery_base;
         static double recovery_exp;
 
-        uint id;
+        uint64_t id;
         double posture;
        
         std::vector<Insurer>* insurers;
@@ -33,8 +33,8 @@ class Defender : public Player {
         PolicyType policy;
 
         Defender(int id_in, Params &p, std::vector<Insurer> &insurers);
-        void gain(uint32_t gain) override;
-        void lose(uint32_t loss) override;
+        void gain(int64_t gain) override;
+        void lose(int64_t loss) override;
 
         void choose_security_strategy();
         void submit_claim(uint32_t loss);
